+++ libraries/ghc-boot-th/GHC/LanguageExtensions/Type.hs
@@ -80,6 +80,7 @@ data Extension
    | DefaultSignatures        -- Allow extra signatures for defmeths
    | DeriveAnyClass           -- Allow deriving any class
    | DeriveLift               -- Allow deriving Lift
+   | DerivingStrategies
 
    | TypeSynonymInstances
    | FlexibleContexts